+++ internal/config/config.go
@@ -198,6 +198,7 @@ type Options struct {
 
 	ExtensionOrder  []string
 	MainFields      []string
+	AbsNodePaths    []string // The "NODE_PATH" variable from Node.js
 	ExternalModules ExternalModules
 
 	AbsOutputFile      string