+++ internal/fs/iswin_other.go
@@ -3,6 +3,6 @@
 
 package fs
 
-func checkIfWindows() bool {
+func CheckIfWindows() bool {
 	return false
 }