+++ pkg/cli/cli_impl.go
@@ -575,6 +575,22 @@ func runImpl(osArgs []string) int {
 			}
 		}
 
+		// Read the "NODE_PATH" from the environment. This is part of node's
+		// module resolution algorithm. Documentation for this can be found here:
+		// https://nodejs.org/api/modules.html#modules_loading_from_the_global_folders
+		for _, key := range os.Environ() {
+			if strings.HasPrefix(key, "NODE_PATH=") {
+				value := key[len("NODE_PATH="):]
+				separator := ":"
+				if fs.CheckIfWindows() {
+					// On Windows, NODE_PATH is delimited by semicolons instead of colons
+					separator = ";"
+				}
+				buildOptions.NodePaths = strings.Split(value, separator)
+				break
+			}
+		}
+
 		// Read from stdin when there are no entry points
 		if len(buildOptions.EntryPoints) == 0 {
 			if buildOptions.Stdin == nil {