+++ scripts/js-api-tests.js
@@ -72,6 +72,30 @@ let buildTests = {
     }
   },
 
+  async nodePathsTest({ esbuild, service, testDir }) {
+    for (const toTest of [esbuild, service]) {
+      let srcDir = path.join(testDir, 'src');
+      let pkgDir = path.join(testDir, 'pkg');
+      let outfile = path.join(testDir, 'out.js');
+      let entry = path.join(srcDir, 'entry.js');
+      let other = path.join(pkgDir, 'other.js');
+      fs.mkdirSync(srcDir, { recursive: true });
+      fs.mkdirSync(pkgDir, { recursive: true });
+      fs.writeFileSync(entry, `export {x} from 'other'`);
+      fs.writeFileSync(other, `export let x = 123`);
+
+      await toTest.build({
+        entryPoints: [entry],
+        outfile,
+        bundle: true,
+        nodePaths: [pkgDir],
+        format: 'cjs',
+      })
+
+      assert.strictEqual(require(outfile).x, 123)
+    }
+  },
+
   async es6_to_cjs({ esbuild, testDir }) {
     const input = path.join(testDir, 'in.js')
     const output = path.join(testDir, 'out.js')